requirements.txt
base32hex==1.0.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cipher==0.1
click==8.0.4
colorama==0.4.4
Flask==2.0.3
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
Naked==0.1.31
pyasn1==0.4.8
pycparser==2.21
pycrypto==2.6.1
PyYAML==6.0
requests==2.27.1
rsa==4.8
shellescape==3.8.1
urllib3==1.26.8
Werkzeug==2.0.3
